LayoutTests/fast/images/zoomed-offset-size-expected.txt

PASS getSize(0.01).imageWidth is 250
PASS getSize(0.01).imageHeight is 166
PASS getSize(0.01).wrapWidth is 250
PASS getSize(0.01).wrapHeight is 166
PASS getSize(0.02).imageWidth is 250
PASS getSize(0.02).imageHeight is 166
PASS getSize(0.02).wrapWidth is 250
PASS getSize(0.02).wrapHeight is 166
PASS getSize(0.03).imageWidth is 250
PASS getSize(0.03).imageHeight is 166
PASS getSize(0.03).wrapWidth is 250
PASS getSize(0.03).wrapHeight is 166
PASS getSize(0.04).imageWidth is 250
PASS getSize(0.04).imageHeight is 166
PASS getSize(0.04).wrapWidth is 250
PASS getSize(0.04).wrapHeight is 166
PASS getSize(0.05).imageWidth is 250
PASS getSize(0.05).imageHeight is 166
PASS getSize(0.05).wrapWidth is 250
PASS getSize(0.05).wrapHeight is 166
PASS getSize(0.06).imageWidth is 250
PASS getSize(0.06).imageHeight is 166
PASS getSize(0.06).wrapWidth is 250
PASS getSize(0.06).wrapHeight is 166
PASS getSize(0.07).imageWidth is 250
PASS getSize(0.07).imageHeight is 166
PASS getSize(0.07).wrapWidth is 250
PASS getSize(0.07).wrapHeight is 166
PASS getSize(0.08).imageWidth is 250
PASS getSize(0.08).imageHeight is 166
PASS getSize(0.08).wrapWidth is 250
PASS getSize(0.08).wrapHeight is 166
PASS getSize(0.09).imageWidth is 250
PASS getSize(0.09).imageHeight is 166
PASS getSize(0.09).wrapWidth is 250
PASS getSize(0.09).wrapHeight is 166
PASS getSize(0.1).imageWidth is 250
PASS getSize(0.1).imageHeight is 166
PASS getSize(0.1).wrapWidth is 250
PASS getSize(0.1).wrapHeight is 166
PASS getSize(0.11).imageWidth is 250
PASS getSize(0.11).imageHeight is 166
PASS getSize(0.11).wrapWidth is 250
PASS getSize(0.11).wrapHeight is 166
PASS getSize(0.12).imageWidth is 250
PASS getSize(0.12).imageHeight is 166
PASS getSize(0.12).wrapWidth is 250
PASS getSize(0.12).wrapHeight is 166
PASS getSize(0.13).imageWidth is 250
PASS getSize(0.13).imageHeight is 166
PASS getSize(0.13).wrapWidth is 250
PASS getSize(0.13).wrapHeight is 166
PASS getSize(0.14).imageWidth is 250
PASS getSize(0.14).imageHeight is 166
PASS getSize(0.14).wrapWidth is 250
PASS getSize(0.14).wrapHeight is 166
PASS getSize(0.15).imageWidth is 250
PASS getSize(0.15).imageHeight is 166
PASS getSize(0.15).wrapWidth is 250
PASS getSize(0.15).wrapHeight is 166
PASS getSize(0.16).imageWidth is 250
PASS getSize(0.16).imageHeight is 166
PASS getSize(0.16).wrapWidth is 250
PASS getSize(0.16).wrapHeight is 166
PASS getSize(0.17).imageWidth is 250
PASS getSize(0.17).imageHeight is 166
PASS getSize(0.17).wrapWidth is 250
PASS getSize(0.17).wrapHeight is 166
PASS getSize(0.18).imageWidth is 250
PASS getSize(0.18).imageHeight is 166
PASS getSize(0.18).wrapWidth is 250
PASS getSize(0.18).wrapHeight is 166
PASS getSize(0.19).imageWidth is 250
PASS getSize(0.19).imageHeight is 166
PASS getSize(0.19).wrapWidth is 250
PASS getSize(0.19).wrapHeight is 166
PASS getSize(0.2).imageWidth is 250
PASS getSize(0.2).imageHeight is 166
PASS getSize(0.2).wrapWidth is 250
PASS getSize(0.2).wrapHeight is 166
PASS getSize(0.21).imageWidth is 250
PASS getSize(0.21).imageHeight is 166
PASS getSize(0.21).wrapWidth is 250
PASS getSize(0.21).wrapHeight is 166
PASS getSize(0.22).imageWidth is 250
PASS getSize(0.22).imageHeight is 166
PASS getSize(0.22).wrapWidth is 250
PASS getSize(0.22).wrapHeight is 166
PASS getSize(0.23).imageWidth is 250
PASS getSize(0.23).imageHeight is 166
PASS getSize(0.23).wrapWidth is 250
PASS getSize(0.23).wrapHeight is 166
PASS getSize(0.24).imageWidth is 250
PASS getSize(0.24).imageHeight is 166
PASS getSize(0.24).wrapWidth is 250
PASS getSize(0.24).wrapHeight is 166
PASS getSize(0.25).imageWidth is 250
PASS getSize(0.25).imageHeight is 166
PASS getSize(0.25).wrapWidth is 250
PASS getSize(0.25).wrapHeight is 166
PASS getSize(0.26).imageWidth is 250
PASS getSize(0.26).imageHeight is 166
PASS getSize(0.26).wrapWidth is 250
PASS getSize(0.26).wrapHeight is 166
PASS getSize(0.27).imageWidth is 250
PASS getSize(0.27).imageHeight is 166
PASS getSize(0.27).wrapWidth is 250
PASS getSize(0.27).wrapHeight is 166
PASS getSize(0.28).imageWidth is 250
PASS getSize(0.28).imageHeight is 166
PASS getSize(0.28).wrapWidth is 250
PASS getSize(0.28).wrapHeight is 166
PASS getSize(0.29).imageWidth is 250
PASS getSize(0.29).imageHeight is 166
PASS getSize(0.29).wrapWidth is 250
PASS getSize(0.29).wrapHeight is 166
PASS getSize(0.3).imageWidth is 250
PASS getSize(0.3).imageHeight is 166
PASS getSize(0.3).wrapWidth is 250
PASS getSize(0.3).wrapHeight is 166
PASS getSize(0.31).imageWidth is 250
PASS getSize(0.31).imageHeight is 166
PASS getSize(0.31).wrapWidth is 250
PASS getSize(0.31).wrapHeight is 166
PASS getSize(0.32).imageWidth is 250
PASS getSize(0.32).imageHeight is 166
PASS getSize(0.32).wrapWidth is 250
PASS getSize(0.32).wrapHeight is 166
PASS getSize(0.33).imageWidth is 250
PASS getSize(0.33).imageHeight is 166
PASS getSize(0.33).wrapWidth is 250
PASS getSize(0.33).wrapHeight is 166
PASS getSize(0.34).imageWidth is 250
PASS getSize(0.34).imageHeight is 166
PASS getSize(0.34).wrapWidth is 250
PASS getSize(0.34).wrapHeight is 166
PASS getSize(0.35).imageWidth is 250
PASS getSize(0.35).imageHeight is 166
PASS getSize(0.35).wrapWidth is 250
PASS getSize(0.35).wrapHeight is 166
PASS getSize(0.36).imageWidth is 250
PASS getSize(0.36).imageHeight is 166
PASS getSize(0.36).wrapWidth is 250
PASS getSize(0.36).wrapHeight is 166
PASS getSize(0.37).imageWidth is 250
PASS getSize(0.37).imageHeight is 166
PASS getSize(0.37).wrapWidth is 250
PASS getSize(0.37).wrapHeight is 166
PASS getSize(0.38).imageWidth is 250
PASS getSize(0.38).imageHeight is 166
PASS getSize(0.38).wrapWidth is 250
PASS getSize(0.38).wrapHeight is 166
PASS getSize(0.39).imageWidth is 250
PASS getSize(0.39).imageHeight is 166
PASS getSize(0.39).wrapWidth is 250
PASS getSize(0.39).wrapHeight is 166
PASS getSize(0.4).imageWidth is 250
PASS getSize(0.4).imageHeight is 166
PASS getSize(0.4).wrapWidth is 250
PASS getSize(0.4).wrapHeight is 166
PASS getSize(0.41).imageWidth is 250
PASS getSize(0.41).imageHeight is 166
PASS getSize(0.41).wrapWidth is 250
PASS getSize(0.41).wrapHeight is 166
PASS getSize(0.42).imageWidth is 250
PASS getSize(0.42).imageHeight is 166
PASS getSize(0.42).wrapWidth is 250
PASS getSize(0.42).wrapHeight is 166
PASS getSize(0.43).imageWidth is 250
PASS getSize(0.43).imageHeight is 166
PASS getSize(0.43).wrapWidth is 250
PASS getSize(0.43).wrapHeight is 166
PASS getSize(0.44).imageWidth is 250
PASS getSize(0.44).imageHeight is 166
PASS getSize(0.44).wrapWidth is 250
PASS getSize(0.44).wrapHeight is 166
PASS getSize(0.45).imageWidth is 250
PASS getSize(0.45).imageHeight is 166
PASS getSize(0.45).wrapWidth is 250
PASS getSize(0.45).wrapHeight is 166
PASS getSize(0.46).imageWidth is 250
PASS getSize(0.46).imageHeight is 166
PASS getSize(0.46).wrapWidth is 250
PASS getSize(0.46).wrapHeight is 166
PASS getSize(0.47).imageWidth is 250
PASS getSize(0.47).imageHeight is 166
PASS getSize(0.47).wrapWidth is 250
PASS getSize(0.47).wrapHeight is 166
PASS getSize(0.48).imageWidth is 250
PASS getSize(0.48).imageHeight is 166
PASS getSize(0.48).wrapWidth is 250
PASS getSize(0.48).wrapHeight is 166
PASS getSize(0.49).imageWidth is 250
PASS getSize(0.49).imageHeight is 166
PASS getSize(0.49).wrapWidth is 250
PASS getSize(0.49).wrapHeight is 166
PASS getSize(0.5).imageWidth is 250
PASS getSize(0.5).imageHeight is 166
PASS getSize(0.5).wrapWidth is 250
PASS getSize(0.5).wrapHeight is 166
PASS getSize(0.51).imageWidth is 250
PASS getSize(0.51).imageHeight is 166
PASS getSize(0.51).wrapWidth is 250
PASS getSize(0.51).wrapHeight is 166
PASS getSize(0.52).imageWidth is 250
PASS getSize(0.52).imageHeight is 166
PASS getSize(0.52).wrapWidth is 250
PASS getSize(0.52).wrapHeight is 166
PASS getSize(0.53).imageWidth is 250
PASS getSize(0.53).imageHeight is 166
PASS getSize(0.53).wrapWidth is 250
PASS getSize(0.53).wrapHeight is 166
PASS getSize(0.54).imageWidth is 250
PASS getSize(0.54).imageHeight is 166
PASS getSize(0.54).wrapWidth is 250
PASS getSize(0.54).wrapHeight is 166
PASS getSize(0.55).imageWidth is 250
PASS getSize(0.55).imageHeight is 166
PASS getSize(0.55).wrapWidth is 250
PASS getSize(0.55).wrapHeight is 166
PASS getSize(0.56).imageWidth is 250
PASS getSize(0.56).imageHeight is 166
PASS getSize(0.56).wrapWidth is 250
PASS getSize(0.56).wrapHeight is 166
PASS getSize(0.57).imageWidth is 250
PASS getSize(0.57).imageHeight is 166
PASS getSize(0.57).wrapWidth is 250
PASS getSize(0.57).wrapHeight is 166
PASS getSize(0.58).imageWidth is 250
PASS getSize(0.58).imageHeight is 166
PASS getSize(0.58).wrapWidth is 250
PASS getSize(0.58).wrapHeight is 166
PASS getSize(0.59).imageWidth is 250
PASS getSize(0.59).imageHeight is 166
PASS getSize(0.59).wrapWidth is 250
PASS getSize(0.59).wrapHeight is 166
PASS getSize(0.6).imageWidth is 250
PASS getSize(0.6).imageHeight is 166
PASS getSize(0.6).wrapWidth is 250
PASS getSize(0.6).wrapHeight is 166
PASS getSize(0.61).imageWidth is 250
PASS getSize(0.61).imageHeight is 166
PASS getSize(0.61).wrapWidth is 250
PASS getSize(0.61).wrapHeight is 166
PASS getSize(0.62).imageWidth is 250
PASS getSize(0.62).imageHeight is 166
PASS getSize(0.62).wrapWidth is 250
PASS getSize(0.62).wrapHeight is 166
PASS getSize(0.63).imageWidth is 250
PASS getSize(0.63).imageHeight is 166
PASS getSize(0.63).wrapWidth is 250
PASS getSize(0.63).wrapHeight is 166
PASS getSize(0.64).imageWidth is 250
PASS getSize(0.64).imageHeight is 166
PASS getSize(0.64).wrapWidth is 250
PASS getSize(0.64).wrapHeight is 166
PASS getSize(0.65).imageWidth is 250
PASS getSize(0.65).imageHeight is 166
PASS getSize(0.65).wrapWidth is 250
PASS getSize(0.65).wrapHeight is 166
PASS getSize(0.66).imageWidth is 250
PASS getSize(0.66).imageHeight is 166
PASS getSize(0.66).wrapWidth is 250
PASS getSize(0.66).wrapHeight is 166
PASS getSize(0.67).imageWidth is 250
PASS getSize(0.67).imageHeight is 166
PASS getSize(0.67).wrapWidth is 250
PASS getSize(0.67).wrapHeight is 166
PASS getSize(0.68).imageWidth is 250
PASS getSize(0.68).imageHeight is 166
PASS getSize(0.68).wrapWidth is 250
PASS getSize(0.68).wrapHeight is 166
PASS getSize(0.69).imageWidth is 250
PASS getSize(0.69).imageHeight is 166
PASS getSize(0.69).wrapWidth is 250
PASS getSize(0.69).wrapHeight is 166
PASS getSize(0.7).imageWidth is 250
PASS getSize(0.7).imageHeight is 166
PASS getSize(0.7).wrapWidth is 250
PASS getSize(0.7).wrapHeight is 166
PASS getSize(0.71).imageWidth is 250
PASS getSize(0.71).imageHeight is 166
PASS getSize(0.71).wrapWidth is 250
PASS getSize(0.71).wrapHeight is 166
PASS getSize(0.72).imageWidth is 250
PASS getSize(0.72).imageHeight is 166
PASS getSize(0.72).wrapWidth is 250
PASS getSize(0.72).wrapHeight is 166
PASS getSize(0.73).imageWidth is 250
PASS getSize(0.73).imageHeight is 166
PASS getSize(0.73).wrapWidth is 250
PASS getSize(0.73).wrapHeight is 166
PASS getSize(0.74).imageWidth is 250
PASS getSize(0.74).imageHeight is 166
PASS getSize(0.74).wrapWidth is 250
PASS getSize(0.74).wrapHeight is 166
PASS getSize(0.75).imageWidth is 250
PASS getSize(0.75).imageHeight is 166
PASS getSize(0.75).wrapWidth is 250
PASS getSize(0.75).wrapHeight is 166
PASS getSize(0.76).imageWidth is 250
PASS getSize(0.76).imageHeight is 166
PASS getSize(0.76).wrapWidth is 250
PASS getSize(0.76).wrapHeight is 166
PASS getSize(0.77).imageWidth is 250
PASS getSize(0.77).imageHeight is 166
PASS getSize(0.77).wrapWidth is 250
PASS getSize(0.77).wrapHeight is 166
PASS getSize(0.78).imageWidth is 250
PASS getSize(0.78).imageHeight is 166
PASS getSize(0.78).wrapWidth is 250
PASS getSize(0.78).wrapHeight is 166
PASS getSize(0.79).imageWidth is 250
PASS getSize(0.79).imageHeight is 166
PASS getSize(0.79).wrapWidth is 250
PASS getSize(0.79).wrapHeight is 166
PASS getSize(0.8).imageWidth is 250
PASS getSize(0.8).imageHeight is 166
PASS getSize(0.8).wrapWidth is 250
PASS getSize(0.8).wrapHeight is 166
PASS getSize(0.81).imageWidth is 250
PASS getSize(0.81).imageHeight is 166
PASS getSize(0.81).wrapWidth is 250
PASS getSize(0.81).wrapHeight is 166
PASS getSize(0.82).imageWidth is 250
PASS getSize(0.82).imageHeight is 166
PASS getSize(0.82).wrapWidth is 250
PASS getSize(0.82).wrapHeight is 166
PASS getSize(0.83).imageWidth is 250
PASS getSize(0.83).imageHeight is 166
PASS getSize(0.83).wrapWidth is 250
PASS getSize(0.83).wrapHeight is 166
PASS getSize(0.84).imageWidth is 250
PASS getSize(0.84).imageHeight is 166
PASS getSize(0.84).wrapWidth is 250
PASS getSize(0.84).wrapHeight is 166
PASS getSize(0.85).imageWidth is 250
PASS getSize(0.85).imageHeight is 166
PASS getSize(0.85).wrapWidth is 250
PASS getSize(0.85).wrapHeight is 166
PASS getSize(0.86).imageWidth is 250
PASS getSize(0.86).imageHeight is 166
PASS getSize(0.86).wrapWidth is 250
PASS getSize(0.86).wrapHeight is 166
PASS getSize(0.87).imageWidth is 250
PASS getSize(0.87).imageHeight is 166
PASS getSize(0.87).wrapWidth is 250
PASS getSize(0.87).wrapHeight is 166
PASS getSize(0.88).imageWidth is 250
PASS getSize(0.88).imageHeight is 166
PASS getSize(0.88).wrapWidth is 250
PASS getSize(0.88).wrapHeight is 166
PASS getSize(0.89).imageWidth is 250
PASS getSize(0.89).imageHeight is 166
PASS getSize(0.89).wrapWidth is 250
PASS getSize(0.89).wrapHeight is 166
PASS getSize(0.9).imageWidth is 250
PASS getSize(0.9).imageHeight is 166
PASS getSize(0.9).wrapWidth is 250
PASS getSize(0.9).wrapHeight is 166
PASS getSize(0.91).imageWidth is 250
PASS getSize(0.91).imageHeight is 166
PASS getSize(0.91).wrapWidth is 250
PASS getSize(0.91).wrapHeight is 166
PASS getSize(0.92).imageWidth is 250
PASS getSize(0.92).imageHeight is 166
PASS getSize(0.92).wrapWidth is 250
PASS getSize(0.92).wrapHeight is 166
PASS getSize(0.93).imageWidth is 250
PASS getSize(0.93).imageHeight is 166
PASS getSize(0.93).wrapWidth is 250
PASS getSize(0.93).wrapHeight is 166
PASS getSize(0.94).imageWidth is 250
PASS getSize(0.94).imageHeight is 166
PASS getSize(0.94).wrapWidth is 250
PASS getSize(0.94).wrapHeight is 166
PASS getSize(0.95).imageWidth is 250
PASS getSize(0.95).imageHeight is 166
PASS getSize(0.95).wrapWidth is 250
PASS getSize(0.95).wrapHeight is 166
PASS getSize(0.96).imageWidth is 250
PASS getSize(0.96).imageHeight is 166
PASS getSize(0.96).wrapWidth is 250
PASS getSize(0.96).wrapHeight is 166
PASS getSize(0.97).imageWidth is 250
PASS getSize(0.97).imageHeight is 166
PASS getSize(0.97).wrapWidth is 250
PASS getSize(0.97).wrapHeight is 166
PASS getSize(0.98).imageWidth is 250
PASS getSize(0.98).imageHeight is 166
PASS getSize(0.98).wrapWidth is 250
PASS getSize(0.98).wrapHeight is 166
PASS getSize(0.99).imageWidth is 250
PASS getSize(0.99).imageHeight is 166
PASS getSize(0.99).wrapWidth is 250
PASS getSize(0.99).wrapHeight is 166
PASS getSize(1).imageWidth is 250
PASS getSize(1).imageHeight is 166
PASS getSize(1).wrapWidth is 250
PASS getSize(1).wrapHeight is 166
PASS getSize(1.01).imageWidth is 250
PASS getSize(1.01).imageHeight is 166
PASS getSize(1.01).wrapWidth is 250
PASS getSize(1.01).wrapHeight is 166
PASS getSize(1.02).imageWidth is 250
PASS getSize(1.02).imageHeight is 166
PASS getSize(1.02).wrapWidth is 250
PASS getSize(1.02).wrapHeight is 166
PASS getSize(1.03).imageWidth is 250
PASS getSize(1.03).imageHeight is 166
PASS getSize(1.03).wrapWidth is 250
PASS getSize(1.03).wrapHeight is 166
PASS getSize(1.04).imageWidth is 250
PASS getSize(1.04).imageHeight is 166
PASS getSize(1.04).wrapWidth is 250
PASS getSize(1.04).wrapHeight is 166
PASS getSize(1.05).imageWidth is 250
PASS getSize(1.05).imageHeight is 166
PASS getSize(1.05).wrapWidth is 250
PASS getSize(1.05).wrapHeight is 166
PASS getSize(1.06).imageWidth is 250
PASS getSize(1.06).imageHeight is 166
PASS getSize(1.06).wrapWidth is 250
PASS getSize(1.06).wrapHeight is 166
PASS getSize(1.07).imageWidth is 250
PASS getSize(1.07).imageHeight is 166
PASS getSize(1.07).wrapWidth is 250
PASS getSize(1.07).wrapHeight is 166
PASS getSize(1.08).imageWidth is 250
PASS getSize(1.08).imageHeight is 166
PASS getSize(1.08).wrapWidth is 250
PASS getSize(1.08).wrapHeight is 166
PASS getSize(1.09).imageWidth is 250
PASS getSize(1.09).imageHeight is 166
PASS getSize(1.09).wrapWidth is 250
PASS getSize(1.09).wrapHeight is 166
PASS getSize(1.1).imageWidth is 250
PASS getSize(1.1).imageHeight is 166
PASS getSize(1.1).wrapWidth is 250
PASS getSize(1.1).wrapHeight is 166
PASS getSize(1.11).imageWidth is 250
PASS getSize(1.11).imageHeight is 166
PASS getSize(1.11).wrapWidth is 250
PASS getSize(1.11).wrapHeight is 166
PASS getSize(1.12).imageWidth is 250
PASS getSize(1.12).imageHeight is 166
PASS getSize(1.12).wrapWidth is 250
PASS getSize(1.12).wrapHeight is 166
PASS getSize(1.13).imageWidth is 250
PASS getSize(1.13).imageHeight is 166
PASS getSize(1.13).wrapWidth is 250
PASS getSize(1.13).wrapHeight is 166
PASS getSize(1.14).imageWidth is 250
PASS getSize(1.14).imageHeight is 166
PASS getSize(1.14).wrapWidth is 250
PASS getSize(1.14).wrapHeight is 166
PASS getSize(1.15).imageWidth is 250
PASS getSize(1.15).imageHeight is 166
PASS getSize(1.15).wrapWidth is 250
PASS getSize(1.15).wrapHeight is 166
PASS getSize(1.16).imageWidth is 250
PASS getSize(1.16).imageHeight is 166
PASS getSize(1.16).wrapWidth is 250
PASS getSize(1.16).wrapHeight is 166
PASS getSize(1.17).imageWidth is 250
PASS getSize(1.17).imageHeight is 166
PASS getSize(1.17).wrapWidth is 250
PASS getSize(1.17).wrapHeight is 166
PASS getSize(1.18).imageWidth is 250
PASS getSize(1.18).imageHeight is 166
PASS getSize(1.18).wrapWidth is 250
PASS getSize(1.18).wrapHeight is 166
PASS getSize(1.19).imageWidth is 250
PASS getSize(1.19).imageHeight is 166
PASS getSize(1.19).wrapWidth is 250
PASS getSize(1.19).wrapHeight is 166
PASS getSize(1.2).imageWidth is 250
PASS getSize(1.2).imageHeight is 166
PASS getSize(1.2).wrapWidth is 250
PASS getSize(1.2).wrapHeight is 166
PASS getSize(1.21).imageWidth is 250
PASS getSize(1.21).imageHeight is 166
PASS getSize(1.21).wrapWidth is 250
PASS getSize(1.21).wrapHeight is 166
PASS getSize(1.22).imageWidth is 250
PASS getSize(1.22).imageHeight is 166
PASS getSize(1.22).wrapWidth is 250
PASS getSize(1.22).wrapHeight is 166
PASS getSize(1.23).imageWidth is 250
PASS getSize(1.23).imageHeight is 166
PASS getSize(1.23).wrapWidth is 250
PASS getSize(1.23).wrapHeight is 166
PASS getSize(1.24).imageWidth is 250
PASS getSize(1.24).imageHeight is 166
PASS getSize(1.24).wrapWidth is 250
PASS getSize(1.24).wrapHeight is 166
PASS getSize(1.25).imageWidth is 250
PASS getSize(1.25).imageHeight is 166
PASS getSize(1.25).wrapWidth is 250
PASS getSize(1.25).wrapHeight is 166
PASS getSize(1.26).imageWidth is 250
PASS getSize(1.26).imageHeight is 166
PASS getSize(1.26).wrapWidth is 250
PASS getSize(1.26).wrapHeight is 166
PASS getSize(1.27).imageWidth is 250
PASS getSize(1.27).imageHeight is 166
PASS getSize(1.27).wrapWidth is 250
PASS getSize(1.27).wrapHeight is 166
PASS getSize(1.28).imageWidth is 250
PASS getSize(1.28).imageHeight is 166
PASS getSize(1.28).wrapWidth is 250
PASS getSize(1.28).wrapHeight is 166
PASS getSize(1.29).imageWidth is 250
PASS getSize(1.29).imageHeight is 166
PASS getSize(1.29).wrapWidth is 250
PASS getSize(1.29).wrapHeight is 166
PASS getSize(1.3).imageWidth is 250
PASS getSize(1.3).imageHeight is 166
PASS getSize(1.3).wrapWidth is 250
PASS getSize(1.3).wrapHeight is 166
PASS getSize(1.31).imageWidth is 250
PASS getSize(1.31).imageHeight is 166
PASS getSize(1.31).wrapWidth is 250
PASS getSize(1.31).wrapHeight is 166
PASS getSize(1.32).imageWidth is 250
PASS getSize(1.32).imageHeight is 166
PASS getSize(1.32).wrapWidth is 250
PASS getSize(1.32).wrapHeight is 166
PASS getSize(1.33).imageWidth is 250
PASS getSize(1.33).imageHeight is 166
PASS getSize(1.33).wrapWidth is 250
PASS getSize(1.33).wrapHeight is 166
PASS getSize(1.34).imageWidth is 250
PASS getSize(1.34).imageHeight is 166
PASS getSize(1.34).wrapWidth is 250
PASS getSize(1.34).wrapHeight is 166
PASS getSize(1.35).imageWidth is 250
PASS getSize(1.35).imageHeight is 166
PASS getSize(1.35).wrapWidth is 250
PASS getSize(1.35).wrapHeight is 166
PASS getSize(1.36).imageWidth is 250
PASS getSize(1.36).imageHeight is 166
PASS getSize(1.36).wrapWidth is 250
PASS getSize(1.36).wrapHeight is 166
PASS getSize(1.37).imageWidth is 250
PASS getSize(1.37).imageHeight is 166
PASS getSize(1.37).wrapWidth is 250
PASS getSize(1.37).wrapHeight is 166
PASS getSize(1.38).imageWidth is 250
PASS getSize(1.38).imageHeight is 166
PASS getSize(1.38).wrapWidth is 250
PASS getSize(1.38).wrapHeight is 166
PASS getSize(1.39).imageWidth is 250
PASS getSize(1.39).imageHeight is 166
PASS getSize(1.39).wrapWidth is 250
PASS getSize(1.39).wrapHeight is 166
PASS getSize(1.4).imageWidth is 250
PASS getSize(1.4).imageHeight is 166
PASS getSize(1.4).wrapWidth is 250
PASS getSize(1.4).wrapHeight is 166
PASS getSize(1.41).imageWidth is 250
PASS getSize(1.41).imageHeight is 166
PASS getSize(1.41).wrapWidth is 250
PASS getSize(1.41).wrapHeight is 166
PASS getSize(1.42).imageWidth is 250
PASS getSize(1.42).imageHeight is 166
PASS getSize(1.42).wrapWidth is 250
PASS getSize(1.42).wrapHeight is 166
PASS getSize(1.43).imageWidth is 250
PASS getSize(1.43).imageHeight is 166
PASS getSize(1.43).wrapWidth is 250
PASS getSize(1.43).wrapHeight is 166
PASS getSize(1.44).imageWidth is 250
PASS getSize(1.44).imageHeight is 166
PASS getSize(1.44).wrapWidth is 250
PASS getSize(1.44).wrapHeight is 166
PASS getSize(1.45).imageWidth is 250
PASS getSize(1.45).imageHeight is 166
PASS getSize(1.45).wrapWidth is 250
PASS getSize(1.45).wrapHeight is 166
PASS getSize(1.46).imageWidth is 250
PASS getSize(1.46).imageHeight is 166
PASS getSize(1.46).wrapWidth is 250
PASS getSize(1.46).wrapHeight is 166
PASS getSize(1.47).imageWidth is 250
PASS getSize(1.47).imageHeight is 166
PASS getSize(1.47).wrapWidth is 250
PASS getSize(1.47).wrapHeight is 166
PASS getSize(1.48).imageWidth is 250
PASS getSize(1.48).imageHeight is 166
PASS getSize(1.48).wrapWidth is 250
PASS getSize(1.48).wrapHeight is 166
PASS getSize(1.49).imageWidth is 250
PASS getSize(1.49).imageHeight is 166
PASS getSize(1.49).wrapWidth is 250
PASS getSize(1.49).wrapHeight is 166
PASS getSize(1.5).imageWidth is 250
PASS getSize(1.5).imageHeight is 166
PASS getSize(1.5).wrapWidth is 250
PASS getSize(1.5).wrapHeight is 166
PASS getSize(1.51).imageWidth is 250
PASS getSize(1.51).imageHeight is 166
PASS getSize(1.51).wrapWidth is 250
PASS getSize(1.51).wrapHeight is 166
PASS getSize(1.52).imageWidth is 250
PASS getSize(1.52).imageHeight is 166
PASS getSize(1.52).wrapWidth is 250
PASS getSize(1.52).wrapHeight is 166
PASS getSize(1.53).imageWidth is 250
PASS getSize(1.53).imageHeight is 166
PASS getSize(1.53).wrapWidth is 250
PASS getSize(1.53).wrapHeight is 166
PASS getSize(1.54).imageWidth is 250
PASS getSize(1.54).imageHeight is 166
PASS getSize(1.54).wrapWidth is 250
PASS getSize(1.54).wrapHeight is 166
PASS getSize(1.55).imageWidth is 250
PASS getSize(1.55).imageHeight is 166
PASS getSize(1.55).wrapWidth is 250
PASS getSize(1.55).wrapHeight is 166
PASS getSize(1.56).imageWidth is 250
PASS getSize(1.56).imageHeight is 166
PASS getSize(1.56).wrapWidth is 250
PASS getSize(1.56).wrapHeight is 166
PASS getSize(1.57).imageWidth is 250
PASS getSize(1.57).imageHeight is 166
PASS getSize(1.57).wrapWidth is 250
PASS getSize(1.57).wrapHeight is 166
PASS getSize(1.58).imageWidth is 250
PASS getSize(1.58).imageHeight is 166
PASS getSize(1.58).wrapWidth is 250
PASS getSize(1.58).wrapHeight is 166
PASS getSize(1.59).imageWidth is 250
PASS getSize(1.59).imageHeight is 166
PASS getSize(1.59).wrapWidth is 250
PASS getSize(1.59).wrapHeight is 166
PASS getSize(1.6).imageWidth is 250
PASS getSize(1.6).imageHeight is 166
PASS getSize(1.6).wrapWidth is 250
PASS getSize(1.6).wrapHeight is 166
PASS getSize(1.61).imageWidth is 250
PASS getSize(1.61).imageHeight is 166
PASS getSize(1.61).wrapWidth is 250
PASS getSize(1.61).wrapHeight is 166
PASS getSize(1.62).imageWidth is 250
PASS getSize(1.62).imageHeight is 166
PASS getSize(1.62).wrapWidth is 250
PASS getSize(1.62).wrapHeight is 166
PASS getSize(1.63).imageWidth is 250
PASS getSize(1.63).imageHeight is 166
PASS getSize(1.63).wrapWidth is 250
PASS getSize(1.63).wrapHeight is 166
PASS getSize(1.64).imageWidth is 250
PASS getSize(1.64).imageHeight is 166
PASS getSize(1.64).wrapWidth is 250
PASS getSize(1.64).wrapHeight is 166
PASS getSize(1.65).imageWidth is 250
PASS getSize(1.65).imageHeight is 166
PASS getSize(1.65).wrapWidth is 250
PASS getSize(1.65).wrapHeight is 166
PASS getSize(1.66).imageWidth is 250
PASS getSize(1.66).imageHeight is 166
PASS getSize(1.66).wrapWidth is 250
PASS getSize(1.66).wrapHeight is 166
PASS getSize(1.67).imageWidth is 250
PASS getSize(1.67).imageHeight is 166
PASS getSize(1.67).wrapWidth is 250
PASS getSize(1.67).wrapHeight is 166
PASS getSize(1.68).imageWidth is 250
PASS getSize(1.68).imageHeight is 166
PASS getSize(1.68).wrapWidth is 250
PASS getSize(1.68).wrapHeight is 166
PASS getSize(1.69).imageWidth is 250
PASS getSize(1.69).imageHeight is 166
PASS getSize(1.69).wrapWidth is 250
PASS getSize(1.69).wrapHeight is 166
PASS getSize(1.7).imageWidth is 250
PASS getSize(1.7).imageHeight is 166
PASS getSize(1.7).wrapWidth is 250
PASS getSize(1.7).wrapHeight is 166
PASS getSize(1.71).imageWidth is 250
PASS getSize(1.71).imageHeight is 166
PASS getSize(1.71).wrapWidth is 250
PASS getSize(1.71).wrapHeight is 166
PASS getSize(1.72).imageWidth is 250
PASS getSize(1.72).imageHeight is 166
PASS getSize(1.72).wrapWidth is 250
PASS getSize(1.72).wrapHeight is 166
PASS getSize(1.73).imageWidth is 250
PASS getSize(1.73).imageHeight is 166
PASS getSize(1.73).wrapWidth is 250
PASS getSize(1.73).wrapHeight is 166
PASS getSize(1.74).imageWidth is 250
PASS getSize(1.74).imageHeight is 166
PASS getSize(1.74).wrapWidth is 250
PASS getSize(1.74).wrapHeight is 166
PASS getSize(1.75).imageWidth is 250
PASS getSize(1.75).imageHeight is 166
PASS getSize(1.75).wrapWidth is 250
PASS getSize(1.75).wrapHeight is 166
PASS getSize(1.76).imageWidth is 250
PASS getSize(1.76).imageHeight is 166
PASS getSize(1.76).wrapWidth is 250
PASS getSize(1.76).wrapHeight is 166
PASS getSize(1.77).imageWidth is 250
PASS getSize(1.77).imageHeight is 166
PASS getSize(1.77).wrapWidth is 250
PASS getSize(1.77).wrapHeight is 166
PASS getSize(1.78).imageWidth is 250
PASS getSize(1.78).imageHeight is 166
PASS getSize(1.78).wrapWidth is 250
PASS getSize(1.78).wrapHeight is 166
PASS getSize(1.79).imageWidth is 250
PASS getSize(1.79).imageHeight is 166
PASS getSize(1.79).wrapWidth is 250
PASS getSize(1.79).wrapHeight is 166
PASS getSize(1.8).imageWidth is 250
PASS getSize(1.8).imageHeight is 166
PASS getSize(1.8).wrapWidth is 250
PASS getSize(1.8).wrapHeight is 166
PASS getSize(1.81).imageWidth is 250
PASS getSize(1.81).imageHeight is 166
PASS getSize(1.81).wrapWidth is 250
PASS getSize(1.81).wrapHeight is 166
PASS getSize(1.82).imageWidth is 250
PASS getSize(1.82).imageHeight is 166
PASS getSize(1.82).wrapWidth is 250
PASS getSize(1.82).wrapHeight is 166
PASS getSize(1.83).imageWidth is 250
PASS getSize(1.83).imageHeight is 166
PASS getSize(1.83).wrapWidth is 250
PASS getSize(1.83).wrapHeight is 166
PASS getSize(1.84).imageWidth is 250
PASS getSize(1.84).imageHeight is 166
PASS getSize(1.84).wrapWidth is 250
PASS getSize(1.84).wrapHeight is 166
PASS getSize(1.85).imageWidth is 250
PASS getSize(1.85).imageHeight is 166
PASS getSize(1.85).wrapWidth is 250
PASS getSize(1.85).wrapHeight is 166
PASS getSize(1.86).imageWidth is 250
PASS getSize(1.86).imageHeight is 166
PASS getSize(1.86).wrapWidth is 250
PASS getSize(1.86).wrapHeight is 166
PASS getSize(1.87).imageWidth is 250
PASS getSize(1.87).imageHeight is 166
PASS getSize(1.87).wrapWidth is 250
PASS getSize(1.87).wrapHeight is 166
PASS getSize(1.88).imageWidth is 250
PASS getSize(1.88).imageHeight is 166
PASS getSize(1.88).wrapWidth is 250
PASS getSize(1.88).wrapHeight is 166
PASS getSize(1.89).imageWidth is 250
PASS getSize(1.89).imageHeight is 166
PASS getSize(1.89).wrapWidth is 250
PASS getSize(1.89).wrapHeight is 166
PASS getSize(1.9).imageWidth is 250
PASS getSize(1.9).imageHeight is 166
PASS getSize(1.9).wrapWidth is 250
PASS getSize(1.9).wrapHeight is 166
PASS getSize(1.91).imageWidth is 250
PASS getSize(1.91).imageHeight is 166
PASS getSize(1.91).wrapWidth is 250
PASS getSize(1.91).wrapHeight is 166
PASS getSize(1.92).imageWidth is 250
PASS getSize(1.92).imageHeight is 166
PASS getSize(1.92).wrapWidth is 250
PASS getSize(1.92).wrapHeight is 166
PASS getSize(1.93).imageWidth is 250
PASS getSize(1.93).imageHeight is 166
PASS getSize(1.93).wrapWidth is 250
PASS getSize(1.93).wrapHeight is 166
PASS getSize(1.94).imageWidth is 250
PASS getSize(1.94).imageHeight is 166
PASS getSize(1.94).wrapWidth is 250
PASS getSize(1.94).wrapHeight is 166
PASS getSize(1.95).imageWidth is 250
PASS getSize(1.95).imageHeight is 166
PASS getSize(1.95).wrapWidth is 250
PASS getSize(1.95).wrapHeight is 166
PASS getSize(1.96).imageWidth is 250
PASS getSize(1.96).imageHeight is 166
PASS getSize(1.96).wrapWidth is 250
PASS getSize(1.96).wrapHeight is 166
PASS getSize(1.97).imageWidth is 250
PASS getSize(1.97).imageHeight is 166
PASS getSize(1.97).wrapWidth is 250
PASS getSize(1.97).wrapHeight is 166
PASS getSize(1.98).imageWidth is 250
PASS getSize(1.98).imageHeight is 166
PASS getSize(1.98).wrapWidth is 250
PASS getSize(1.98).wrapHeight is 166
PASS getSize(1.99).imageWidth is 250
PASS getSize(1.99).imageHeight is 166
PASS getSize(1.99).wrapWidth is 250
PASS getSize(1.99).wrapHeight is 166
PASS getSize(2).imageWidth is 250
PASS getSize(2).imageHeight is 166
PASS getSize(2).wrapWidth is 250
PASS getSize(2).wrapHeight is 166
PASS getSize(2.01).imageWidth is 250
PASS getSize(2.01).imageHeight is 166
PASS getSize(2.01).wrapWidth is 250
PASS getSize(2.01).wrapHeight is 166
PASS getSize(2.02).imageWidth is 250
PASS getSize(2.02).imageHeight is 166
PASS getSize(2.02).wrapWidth is 250
PASS getSize(2.02).wrapHeight is 166
PASS getSize(2.03).imageWidth is 250
PASS getSize(2.03).imageHeight is 166
PASS getSize(2.03).wrapWidth is 250
PASS getSize(2.03).wrapHeight is 166
PASS getSize(2.04).imageWidth is 250
PASS getSize(2.04).imageHeight is 166
PASS getSize(2.04).wrapWidth is 250
PASS getSize(2.04).wrapHeight is 166
PASS getSize(2.05).imageWidth is 250
PASS getSize(2.05).imageHeight is 166
PASS getSize(2.05).wrapWidth is 250
PASS getSize(2.05).wrapHeight is 166
PASS getSize(2.06).imageWidth is 250
PASS getSize(2.06).imageHeight is 166
PASS getSize(2.06).wrapWidth is 250
PASS getSize(2.06).wrapHeight is 166
PASS getSize(2.07).imageWidth is 250
PASS getSize(2.07).imageHeight is 166
PASS getSize(2.07).wrapWidth is 250
PASS getSize(2.07).wrapHeight is 166
PASS getSize(2.08).imageWidth is 250
PASS getSize(2.08).imageHeight is 166
PASS getSize(2.08).wrapWidth is 250
PASS getSize(2.08).wrapHeight is 166
PASS getSize(2.09).imageWidth is 250
PASS getSize(2.09).imageHeight is 166
PASS getSize(2.09).wrapWidth is 250
PASS getSize(2.09).wrapHeight is 166
PASS getSize(2.1).imageWidth is 250
PASS getSize(2.1).imageHeight is 166
PASS getSize(2.1).wrapWidth is 250
PASS getSize(2.1).wrapHeight is 166
PASS getSize(2.11).imageWidth is 250
PASS getSize(2.11).imageHeight is 166
PASS getSize(2.11).wrapWidth is 250
PASS getSize(2.11).wrapHeight is 166
PASS getSize(2.12).imageWidth is 250
PASS getSize(2.12).imageHeight is 166
PASS getSize(2.12).wrapWidth is 250
PASS getSize(2.12).wrapHeight is 166
PASS getSize(2.13).imageWidth is 250
PASS getSize(2.13).imageHeight is 166
PASS getSize(2.13).wrapWidth is 250
PASS getSize(2.13).wrapHeight is 166
PASS getSize(2.14).imageWidth is 250
PASS getSize(2.14).imageHeight is 166
PASS getSize(2.14).wrapWidth is 250
PASS getSize(2.14).wrapHeight is 166
PASS getSize(2.15).imageWidth is 250
PASS getSize(2.15).imageHeight is 166
PASS getSize(2.15).wrapWidth is 250
PASS getSize(2.15).wrapHeight is 166
PASS getSize(2.16).imageWidth is 250
PASS getSize(2.16).imageHeight is 166
PASS getSize(2.16).wrapWidth is 250
PASS getSize(2.16).wrapHeight is 166
PASS getSize(2.17).imageWidth is 250
PASS getSize(2.17).imageHeight is 166
PASS getSize(2.17).wrapWidth is 250
PASS getSize(2.17).wrapHeight is 166
PASS getSize(2.18).imageWidth is 250
PASS getSize(2.18).imageHeight is 166
PASS getSize(2.18).wrapWidth is 250
PASS getSize(2.18).wrapHeight is 166
PASS getSize(2.19).imageWidth is 250
PASS getSize(2.19).imageHeight is 166
PASS getSize(2.19).wrapWidth is 250
PASS getSize(2.19).wrapHeight is 166
PASS getSize(2.2).imageWidth is 250
PASS getSize(2.2).imageHeight is 166
PASS getSize(2.2).wrapWidth is 250
PASS getSize(2.2).wrapHeight is 166
PASS getSize(2.21).imageWidth is 250
PASS getSize(2.21).imageHeight is 166
PASS getSize(2.21).wrapWidth is 250
PASS getSize(2.21).wrapHeight is 166
PASS getSize(2.22).imageWidth is 250
PASS getSize(2.22).imageHeight is 166
PASS getSize(2.22).wrapWidth is 250
PASS getSize(2.22).wrapHeight is 166
PASS getSize(2.23).imageWidth is 250
PASS getSize(2.23).imageHeight is 166
PASS getSize(2.23).wrapWidth is 250
PASS getSize(2.23).wrapHeight is 166
PASS getSize(2.24).imageWidth is 250
PASS getSize(2.24).imageHeight is 166
PASS getSize(2.24).wrapWidth is 250
PASS getSize(2.24).wrapHeight is 166
PASS getSize(2.25).imageWidth is 250
PASS getSize(2.25).imageHeight is 166
PASS getSize(2.25).wrapWidth is 250
PASS getSize(2.25).wrapHeight is 166
PASS getSize(2.26).imageWidth is 250
PASS getSize(2.26).imageHeight is 166
PASS getSize(2.26).wrapWidth is 250
PASS getSize(2.26).wrapHeight is 166
PASS getSize(2.27).imageWidth is 250
PASS getSize(2.27).imageHeight is 166
PASS getSize(2.27).wrapWidth is 250
PASS getSize(2.27).wrapHeight is 166
PASS getSize(2.28).imageWidth is 250
PASS getSize(2.28).imageHeight is 166
PASS getSize(2.28).wrapWidth is 250
PASS getSize(2.28).wrapHeight is 166
PASS getSize(2.29).imageWidth is 250
PASS getSize(2.29).imageHeight is 166
PASS getSize(2.29).wrapWidth is 250
PASS getSize(2.29).wrapHeight is 166
PASS getSize(2.3).imageWidth is 250
PASS getSize(2.3).imageHeight is 166
PASS getSize(2.3).wrapWidth is 250
PASS getSize(2.3).wrapHeight is 166
PASS getSize(2.31).imageWidth is 250
PASS getSize(2.31).imageHeight is 166
PASS getSize(2.31).wrapWidth is 250
PASS getSize(2.31).wrapHeight is 166
PASS getSize(2.32).imageWidth is 250
PASS getSize(2.32).imageHeight is 166
PASS getSize(2.32).wrapWidth is 250
PASS getSize(2.32).wrapHeight is 166
PASS getSize(2.33).imageWidth is 250
PASS getSize(2.33).imageHeight is 166
PASS getSize(2.33).wrapWidth is 250
PASS getSize(2.33).wrapHeight is 166
PASS getSize(2.34).imageWidth is 250
PASS getSize(2.34).imageHeight is 166
PASS getSize(2.34).wrapWidth is 250
PASS getSize(2.34).wrapHeight is 166
PASS getSize(2.35).imageWidth is 250
PASS getSize(2.35).imageHeight is 166
PASS getSize(2.35).wrapWidth is 250
PASS getSize(2.35).wrapHeight is 166
PASS getSize(2.36).imageWidth is 250
PASS getSize(2.36).imageHeight is 166
PASS getSize(2.36).wrapWidth is 250
PASS getSize(2.36).wrapHeight is 166
PASS getSize(2.37).imageWidth is 250
PASS getSize(2.37).imageHeight is 166
PASS getSize(2.37).wrapWidth is 250
PASS getSize(2.37).wrapHeight is 166
PASS getSize(2.38).imageWidth is 250
PASS getSize(2.38).imageHeight is 166
PASS getSize(2.38).wrapWidth is 250
PASS getSize(2.38).wrapHeight is 166
PASS getSize(2.39).imageWidth is 250
PASS getSize(2.39).imageHeight is 166
PASS getSize(2.39).wrapWidth is 250
PASS getSize(2.39).wrapHeight is 166
PASS getSize(2.4).imageWidth is 250
PASS getSize(2.4).imageHeight is 166
PASS getSize(2.4).wrapWidth is 250
PASS getSize(2.4).wrapHeight is 166
PASS getSize(2.41).imageWidth is 250
PASS getSize(2.41).imageHeight is 166
PASS getSize(2.41).wrapWidth is 250
PASS getSize(2.41).wrapHeight is 166
PASS getSize(2.42).imageWidth is 250
PASS getSize(2.42).imageHeight is 166
PASS getSize(2.42).wrapWidth is 250
PASS getSize(2.42).wrapHeight is 166
PASS getSize(2.43).imageWidth is 250
PASS getSize(2.43).imageHeight is 166
PASS getSize(2.43).wrapWidth is 250
PASS getSize(2.43).wrapHeight is 166
PASS getSize(2.44).imageWidth is 250
PASS getSize(2.44).imageHeight is 166
PASS getSize(2.44).wrapWidth is 250
PASS getSize(2.44).wrapHeight is 166
PASS getSize(2.45).imageWidth is 250
PASS getSize(2.45).imageHeight is 166
PASS getSize(2.45).wrapWidth is 250
PASS getSize(2.45).wrapHeight is 166
PASS getSize(2.46).imageWidth is 250
PASS getSize(2.46).imageHeight is 166
PASS getSize(2.46).wrapWidth is 250
PASS getSize(2.46).wrapHeight is 166
PASS getSize(2.47).imageWidth is 250
PASS getSize(2.47).imageHeight is 166
PASS getSize(2.47).wrapWidth is 250
PASS getSize(2.47).wrapHeight is 166
PASS getSize(2.48).imageWidth is 250
PASS getSize(2.48).imageHeight is 166
PASS getSize(2.48).wrapWidth is 250
PASS getSize(2.48).wrapHeight is 166
PASS getSize(2.49).imageWidth is 250
PASS getSize(2.49).imageHeight is 166
PASS getSize(2.49).wrapWidth is 250
PASS getSize(2.49).wrapHeight is 166
PASS getSize(2.5).imageWidth is 250
PASS getSize(2.5).imageHeight is 166
PASS getSize(2.5).wrapWidth is 250
PASS getSize(2.5).wrapHeight is 166
PASS getSize(2.51).imageWidth is 250
PASS getSize(2.51).imageHeight is 166
PASS getSize(2.51).wrapWidth is 250
PASS getSize(2.51).wrapHeight is 166
PASS getSize(2.52).imageWidth is 250
PASS getSize(2.52).imageHeight is 166
PASS getSize(2.52).wrapWidth is 250
PASS getSize(2.52).wrapHeight is 166
PASS getSize(2.53).imageWidth is 250
PASS getSize(2.53).imageHeight is 166
PASS getSize(2.53).wrapWidth is 250
PASS getSize(2.53).wrapHeight is 166
PASS getSize(2.54).imageWidth is 250
PASS getSize(2.54).imageHeight is 166
PASS getSize(2.54).wrapWidth is 250
PASS getSize(2.54).wrapHeight is 166
PASS getSize(2.55).imageWidth is 250
PASS getSize(2.55).imageHeight is 166
PASS getSize(2.55).wrapWidth is 250
PASS getSize(2.55).wrapHeight is 166
PASS getSize(2.56).imageWidth is 250
PASS getSize(2.56).imageHeight is 166
PASS getSize(2.56).wrapWidth is 250
PASS getSize(2.56).wrapHeight is 166
PASS getSize(2.57).imageWidth is 250
PASS getSize(2.57).imageHeight is 166
PASS getSize(2.57).wrapWidth is 250
PASS getSize(2.57).wrapHeight is 166
PASS getSize(2.58).imageWidth is 250
PASS getSize(2.58).imageHeight is 166
PASS getSize(2.58).wrapWidth is 250
PASS getSize(2.58).wrapHeight is 166
PASS getSize(2.59).imageWidth is 250
PASS getSize(2.59).imageHeight is 166
PASS getSize(2.59).wrapWidth is 250
PASS getSize(2.59).wrapHeight is 166
PASS getSize(2.6).imageWidth is 250
PASS getSize(2.6).imageHeight is 166
PASS getSize(2.6).wrapWidth is 250
PASS getSize(2.6).wrapHeight is 166
PASS getSize(2.61).imageWidth is 250
PASS getSize(2.61).imageHeight is 166
PASS getSize(2.61).wrapWidth is 250
PASS getSize(2.61).wrapHeight is 166
PASS getSize(2.62).imageWidth is 250
PASS getSize(2.62).imageHeight is 166
PASS getSize(2.62).wrapWidth is 250
PASS getSize(2.62).wrapHeight is 166
PASS getSize(2.63).imageWidth is 250
PASS getSize(2.63).imageHeight is 166
PASS getSize(2.63).wrapWidth is 250
PASS getSize(2.63).wrapHeight is 166
PASS getSize(2.64).imageWidth is 250
PASS getSize(2.64).imageHeight is 166
PASS getSize(2.64).wrapWidth is 250
PASS getSize(2.64).wrapHeight is 166
PASS getSize(2.65).imageWidth is 250
PASS getSize(2.65).imageHeight is 166
PASS getSize(2.65).wrapWidth is 250
PASS getSize(2.65).wrapHeight is 166
PASS getSize(2.66).imageWidth is 250
PASS getSize(2.66).imageHeight is 166
PASS getSize(2.66).wrapWidth is 250
PASS getSize(2.66).wrapHeight is 166
PASS getSize(2.67).imageWidth is 250
PASS getSize(2.67).imageHeight is 166
PASS getSize(2.67).wrapWidth is 250
PASS getSize(2.67).wrapHeight is 166
PASS getSize(2.68).imageWidth is 250
PASS getSize(2.68).imageHeight is 166
PASS getSize(2.68).wrapWidth is 250
PASS getSize(2.68).wrapHeight is 166
PASS getSize(2.69).imageWidth is 250
PASS getSize(2.69).imageHeight is 166
PASS getSize(2.69).wrapWidth is 250
PASS getSize(2.69).wrapHeight is 166
PASS getSize(2.7).imageWidth is 250
PASS getSize(2.7).imageHeight is 166
PASS getSize(2.7).wrapWidth is 250
PASS getSize(2.7).wrapHeight is 166
PASS getSize(2.71).imageWidth is 250
PASS getSize(2.71).imageHeight is 166
PASS getSize(2.71).wrapWidth is 250
PASS getSize(2.71).wrapHeight is 166
PASS getSize(2.72).imageWidth is 250
PASS getSize(2.72).imageHeight is 166
PASS getSize(2.72).wrapWidth is 250
PASS getSize(2.72).wrapHeight is 166
PASS getSize(2.73).imageWidth is 250
PASS getSize(2.73).imageHeight is 166
PASS getSize(2.73).wrapWidth is 250
PASS getSize(2.73).wrapHeight is 166
PASS getSize(2.74).imageWidth is 250
PASS getSize(2.74).imageHeight is 166
PASS getSize(2.74).wrapWidth is 250
PASS getSize(2.74).wrapHeight is 166
PASS getSize(2.75).imageWidth is 250
PASS getSize(2.75).imageHeight is 166
PASS getSize(2.75).wrapWidth is 250
PASS getSize(2.75).wrapHeight is 166
PASS getSize(2.76).imageWidth is 250
PASS getSize(2.76).imageHeight is 166
PASS getSize(2.76).wrapWidth is 250
PASS getSize(2.76).wrapHeight is 166
PASS getSize(2.77).imageWidth is 250
PASS getSize(2.77).imageHeight is 166
PASS getSize(2.77).wrapWidth is 250
PASS getSize(2.77).wrapHeight is 166
PASS getSize(2.78).imageWidth is 250
PASS getSize(2.78).imageHeight is 166
PASS getSize(2.78).wrapWidth is 250
PASS getSize(2.78).wrapHeight is 166
PASS getSize(2.79).imageWidth is 250
PASS getSize(2.79).imageHeight is 166
PASS getSize(2.79).wrapWidth is 250
PASS getSize(2.79).wrapHeight is 166
PASS getSize(2.8).imageWidth is 250
PASS getSize(2.8).imageHeight is 166
PASS getSize(2.8).wrapWidth is 250
PASS getSize(2.8).wrapHeight is 166
PASS getSize(2.81).imageWidth is 250
PASS getSize(2.81).imageHeight is 166
PASS getSize(2.81).wrapWidth is 250
PASS getSize(2.81).wrapHeight is 166
PASS getSize(2.82).imageWidth is 250
PASS getSize(2.82).imageHeight is 166
PASS getSize(2.82).wrapWidth is 250
PASS getSize(2.82).wrapHeight is 166
PASS getSize(2.83).imageWidth is 250
PASS getSize(2.83).imageHeight is 166
PASS getSize(2.83).wrapWidth is 250
PASS getSize(2.83).wrapHeight is 166
PASS getSize(2.84).imageWidth is 250
PASS getSize(2.84).imageHeight is 166
PASS getSize(2.84).wrapWidth is 250
PASS getSize(2.84).wrapHeight is 166
PASS getSize(2.85).imageWidth is 250
PASS getSize(2.85).imageHeight is 166
PASS getSize(2.85).wrapWidth is 250
PASS getSize(2.85).wrapHeight is 166
PASS getSize(2.86).imageWidth is 250
PASS getSize(2.86).imageHeight is 166
PASS getSize(2.86).wrapWidth is 250
PASS getSize(2.86).wrapHeight is 166
PASS getSize(2.87).imageWidth is 250
PASS getSize(2.87).imageHeight is 166
PASS getSize(2.87).wrapWidth is 250
PASS getSize(2.87).wrapHeight is 166
PASS getSize(2.88).imageWidth is 250
PASS getSize(2.88).imageHeight is 166
PASS getSize(2.88).wrapWidth is 250
PASS getSize(2.88).wrapHeight is 166
PASS getSize(2.89).imageWidth is 250
PASS getSize(2.89).imageHeight is 166
PASS getSize(2.89).wrapWidth is 250
PASS getSize(2.89).wrapHeight is 166
PASS getSize(2.9).imageWidth is 250
PASS getSize(2.9).imageHeight is 166
PASS getSize(2.9).wrapWidth is 250
PASS getSize(2.9).wrapHeight is 166
PASS getSize(2.91).imageWidth is 250
PASS getSize(2.91).imageHeight is 166
PASS getSize(2.91).wrapWidth is 250
PASS getSize(2.91).wrapHeight is 166
PASS getSize(2.92).imageWidth is 250
PASS getSize(2.92).imageHeight is 166
PASS getSize(2.92).wrapWidth is 250
PASS getSize(2.92).wrapHeight is 166
PASS getSize(2.93).imageWidth is 250
PASS getSize(2.93).imageHeight is 166
PASS getSize(2.93).wrapWidth is 250
PASS getSize(2.93).wrapHeight is 166
PASS getSize(2.94).imageWidth is 250
PASS getSize(2.94).imageHeight is 166
PASS getSize(2.94).wrapWidth is 250
PASS getSize(2.94).wrapHeight is 166
PASS getSize(2.95).imageWidth is 250
PASS getSize(2.95).imageHeight is 166
PASS getSize(2.95).wrapWidth is 250
PASS getSize(2.95).wrapHeight is 166
PASS getSize(2.96).imageWidth is 250
PASS getSize(2.96).imageHeight is 166
PASS getSize(2.96).wrapWidth is 250
PASS getSize(2.96).wrapHeight is 166
PASS getSize(2.97).imageWidth is 250
PASS getSize(2.97).imageHeight is 166
PASS getSize(2.97).wrapWidth is 250
PASS getSize(2.97).wrapHeight is 166
PASS getSize(2.98).imageWidth is 250
PASS getSize(2.98).imageHeight is 166
PASS getSize(2.98).wrapWidth is 250
PASS getSize(2.98).wrapHeight is 166
PASS getSize(2.99).imageWidth is 250
PASS getSize(2.99).imageHeight is 166
PASS getSize(2.99).wrapWidth is 250
PASS getSize(2.99).wrapHeight is 166
PASS getSize(3).imageWidth is 250
PASS getSize(3).imageHeight is 166
PASS getSize(3).wrapWidth is 250
PASS getSize(3).wrapHeight is 166

Image should return a size of 250x166 regardless of zoom level.
